+++ changelog.md
@@ -16,6 +16,7 @@ All notable changes to this project will be documented in this file.
 * add removal of temp folder
 * create api endpoint for fetching logs
 * create api endpoint for running builder3
+* create api endpoint for clear temp information builder3
 
 # [0.0.74] - 2024-06-10
 * add controllers concept to server